CPU usage change
This is on my computer, so I don't know if anyone else notices or cares, but the script/app would constantly take 1% of my CPU. That doesn't sound like much, but I have well over 100 other processes (499 at this minute, actually) that also want to use CPU and if they each took 1%, the computer wouldn't be usable. In looking at the "CPU time" column in Task Manager, I can see that it rivals my antivirus software for the amount of total CPU it's using. 

See the attached changes. I am not proposing these exact changes get merged in but I figured it was easier to show what I did with a pull request than an Issue... I take out the WinWait and look for the window after a hard-coded (10 seconds) amount of time. Obviously, if we thought this was a good idea, that would be a configurable amount of time. 

After making these changes, the CPU now sits around .03% (as seen in Process Explorer).